Refactor data generation
